+++ docs/pages/resources/changelog.md
@@ -15,6 +15,7 @@ New versions of Shoelace are released as-needed and generally occur when a criti
 ## Next
 
 - Improved performance of `<sl-select>` when using a large number of options [#2318]
+- Improved accessibility of `<sl-carousel>` [#2353]
 - Updated the Japanese translation [#2329]
 
 ## 2.19.1